+++ src/providers/flow/tokenizer/japanese.ts
@@ -1,5 +1,5 @@
 import { TinySegmenter } from 'src/vendor/tiny-segmenter'
-import { Range, Tokenizer } from '../tokenizer'
+import { Range, Tokenizer, TokenizerOptions } from '../tokenizer'
 
 export class JapaneseTokenizer extends Tokenizer {
   // @ts-ignore
@@ -13,5 +13,31 @@ export class JapaneseTokenizer extends Tokenizer {
       .map((t) => t.replace(this.trimPattern, ''))
 
     return { tokens }
+  }
+
+  lastWordFrom(
+    text: string,
+    index: number,
+    options: TokenizerOptions = { normalize: false }
+  ): string | null {
+    const { normalized } = options.normalize
+      ? this.normalizedLine(text, index)
+      : { normalized: text }
+
+    const tokens = this.tokenizer
+      .segment(normalized)
+      .map((t: string) => t.replace(this.trimPattern, ''))
+    const length = tokens.length
+
+    return length > 0 ? tokens[length - 1] : null
+  }
+
+  lastWordStartPos(
+    text: string,
+    index: number,
+    options: TokenizerOptions = { normalize: false }
+  ): number {
+    const lastWord = this.lastWordFrom(text, index, options)
+    return lastWord ? text.length - lastWord.length : 0
   }
 }